+++ sdk/sql/azure-mgmt-sql/generated_tests/test_sql_management_sync_members_operations.py
@@ -26,7 +26,7 @@ class TestSqlManagementSyncMembersOperations(AzureMgmtRecordedTestCase):
             server_name="str",
             database_name="str",
             sync_group_name="str",
-            api_version="2024-05-01-preview",
+            api_version="2024-11-01-preview",
         )
         result = [r for r in response]
         # please add some check logic here by yourself
@@ -41,7 +41,7 @@ class TestSqlManagementSyncMembersOperations(AzureMgmtRecordedTestCase):
             database_name="str",
             sync_group_name="str",
             sync_member_name="str",
-            api_version="2024-05-01-preview",
+            api_version="2024-11-01-preview",
         )
 
         # please add some check logic here by yourself
@@ -73,7 +73,7 @@ class TestSqlManagementSyncMembersOperations(AzureMgmtRecordedTestCase):
                 "usePrivateLinkConnection": bool,
                 "userName": "str",
             },
-            api_version="2024-05-01-preview",
+            api_version="2024-11-01-preview",
         ).result()  # call '.result()' to poll until service return final result
 
         # please add some check logic here by yourself
@@ -88,7 +88,7 @@ class TestSqlManagementSyncMembersOperations(AzureMgmtRecordedTestCase):
             database_name="str",
             sync_group_name="str",
             sync_member_name="str",
-            api_version="2024-05-01-preview",
+            api_version="2024-11-01-preview",
         ).result()  # call '.result()' to poll until service return final result
 
         # please add some check logic here by yourself
@@ -120,7 +120,7 @@ class TestSqlManagementSyncMembersOperations(AzureMgmtRecordedTestCase):
                 "usePrivateLinkConnection": bool,
                 "userName": "str",
             },
-            api_version="2024-05-01-preview",
+            api_version="2024-11-01-preview",
         ).result()  # call '.result()' to poll until service return final result
 
         # please add some check logic here by yourself
@@ -135,7 +135,7 @@ class TestSqlManagementSyncMembersOperations(AzureMgmtRecordedTestCase):
             database_name="str",
             sync_group_name="str",
             sync_member_name="str",
-            api_version="2024-05-01-preview",
+            api_version="2024-11-01-preview",
         ).result()  # call '.result()' to poll until service return final result
 
         # please add some check logic here by yourself
@@ -150,7 +150,7 @@ class TestSqlManagementSyncMembersOperations(AzureMgmtRecordedTestCase):
             database_name="str",
             sync_group_name="str",
             sync_member_name="str",
-            api_version="2024-05-01-preview",
+            api_version="2024-11-01-preview",
         )
         result = [r for r in response]
         # please add some check logic here by yourself