+++ sdk/sql/azure-mgmt-sql/generated_tests/test_sql_management_workload_classifiers_operations_async.py
@@ -27,7 +27,7 @@ class TestSqlManagementWorkloadClassifiersOperationsAsync(AzureMgmtRecordedTestC
             server_name="str",
             database_name="str",
             workload_group_name="str",
-            api_version="2024-05-01-preview",
+            api_version="2024-11-01-preview",
         )
         result = [r async for r in response]
         # please add some check logic here by yourself
@@ -42,7 +42,7 @@ class TestSqlManagementWorkloadClassifiersOperationsAsync(AzureMgmtRecordedTestC
             database_name="str",
             workload_group_name="str",
             workload_classifier_name="str",
-            api_version="2024-05-01-preview",
+            api_version="2024-11-01-preview",
         )
 
         # please add some check logic here by yourself
@@ -69,7 +69,7 @@ class TestSqlManagementWorkloadClassifiersOperationsAsync(AzureMgmtRecordedTestC
                     "startTime": "str",
                     "type": "str",
                 },
-                api_version="2024-05-01-preview",
+                api_version="2024-11-01-preview",
             )
         ).result()  # call '.result()' to poll until service return final result
 
@@ -86,7 +86,7 @@ class TestSqlManagementWorkloadClassifiersOperationsAsync(AzureMgmtRecordedTestC
                 database_name="str",
                 workload_group_name="str",
                 workload_classifier_name="str",
-                api_version="2024-05-01-preview",
+                api_version="2024-11-01-preview",
             )
         ).result()  # call '.result()' to poll until service return final result